// bookscape-ui/src/app/services/models/page-response-book-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BookDto } from '../models/book-dto';
export interface PageResponseBookDto {
  content?: Array<BookDto>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}